modify feature manager implementation, factor out helm service interface
